test(e2e): convert Playwright config and E2E helpers to TypeScript
Also finish the shared contracts test migration.
