+++ tests/e2e/constants.ts
@@ -1,6 +1,2 @@
 /** Shared E2E static server port — keep in sync across Playwright and the local file server. */
-const E2E_FRONTEND_PORT = 4000;
-
-module.exports = {
-  E2E_FRONTEND_PORT,
-};
+export const E2E_FRONTEND_PORT = 4000;
