--- tests/e2e/serve-frontend.js
@@ -1,27 +1,0 @@
-const http = require("node:http");
-const fs = require("node:fs");
-const path = require("node:path");
-const { E2E_FRONTEND_PORT } = require("./constants");
-
-const DIST = path.resolve(__dirname, "../../apps/desktop/dist");
-const PORT = E2E_FRONTEND_PORT;
-
-const MIME = {
-  ".html": "text/html",
-  ".js": "application/javascript",
-  ".css": "text/css",
-};
-
-http.createServer((req, res) => {
-  const file = req.url === "/" ? "index.html" : req.url.slice(1);
-  const filePath = path.join(DIST, file);
-  const ext = path.extname(filePath);
-  try {
-    const content = fs.readFileSync(filePath);
-    res.writeHead(200, { "Content-Type": MIME[ext] || "application/octet-stream" });
-    res.end(content);
-  } catch {
-    res.writeHead(404);
-    res.end("not found");
-  }
-}).listen(PORT, () => console.log(`frontend on http://localhost:${PORT}`));
